ddl: add the db_partition_test
