+++ ddl/db_test.go
@@ -301,8 +301,7 @@ func (s *testDBSuite) TestRenameIndex(c *C) {
 	s.testErrorCode(c, "alter table t rename key k3 to K2", mysql.ErrDupKeyName)
 }
 
-func (s *testDBSuite) testGetTableByName(c *C, db, table string) table.Table {
-	ctx := s.s.(sessionctx.Context)
+func testGetTableByName(c *C, ctx sessionctx.Context, db, table string) table.Table {
 	dom := domain.GetDomain(ctx)
 	// Make sure the table schema is the new schema.
 	err := dom.Reload()
@@ -313,7 +312,8 @@ func (s *testDBSuite) testGetTableByName(c *C, db, table string) table.Table {
 }
 
 func (s *testDBSuite) testGetTable(c *C, name string) table.Table {
-	return s.testGetTableByName(c, s.schemaName, name)
+	ctx := s.s.(sessionctx.Context)
+	return testGetTableByName(c, ctx, s.schemaName, name)
 }
 
 func backgroundExec(s kv.Storage, sql string, done chan error) {
@@ -412,7 +412,7 @@ func (s *testDBSuite) TestCancelAddIndex(c *C) {
 	// After that ddl call d.hook.OnJobUpdated(job), so that we can canceled the job in this test case.
 	ddl.ReorgWaitTimeout = 50 * time.Millisecond
 	var checkErr error
-	hook.OnJobUpdatedExported, c3IdxInfo, checkErr = backgroundExecOnJobUpdatedExported(c, s, hook)
+	hook.OnJobUpdatedExported, c3IdxInfo, checkErr = backgroundExecOnJobUpdatedExported(c, s.store, s.s.(sessionctx.Context), hook)
 	s.dom.DDL().(ddl.DDLForTest).SetHook(hook)
 	done := make(chan error, 1)
 	go backgroundExec(s.store, "create unique index c3_index on t1 (c3)", done)
@@ -566,15 +566,6 @@ func (s *testDBSuite) TestAddAnonymousIndex(c *C) {
 	t = s.testGetTable(c, "t_primary_3")
 	c.Assert(t.Indices()[0].Meta().Name.String(), Equals, "primary_2")
 	c.Assert(t.Indices()[1].Meta().Name.String(), Equals, "primary_3")
-}
-
-// TestModifyColumnAfterAddIndex Issue 5134
-func (s *testDBSuite) TestModifyColumnAfterAddIndex(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use " + s.schemaName)
-	s.mustExec(c, "create table city (city VARCHAR(2) KEY);")
-	s.mustExec(c, "alter table city change column city city varchar(50);")
-	s.mustExec(c, `insert into city values ("abc"), ("abd");`)
 }
 
 func (s *testDBSuite) testAlterLock(c *C) {
@@ -908,32 +899,6 @@ func (s *testDBSuite) showColumns(c *C, tableName string) [][]interface{} {
 	return s.mustQuery(c, fmt.Sprintf("show columns from %s", tableName))
 }
 
-func (s *testDBSuite) TestIssue2293(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use " + s.schemaName)
-	s.tk.MustExec("create table t_issue_2293 (a int)")
-	sql := "alter table t_issue_2293 add b int not null default 'a'"
-	s.testErrorCode(c, sql, tmysql.ErrInvalidDefault)
-	s.tk.MustExec("insert into t_issue_2293 value(1)")
-	s.tk.MustQuery("select * from t_issue_2293").Check(testkit.Rows("1"))
-}
-
-func (s *testDBSuite) TestIssue6101(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use " + s.schemaName)
-	s.tk.MustExec("create table t1 (quantity decimal(2) unsigned);")
-	_, err := s.tk.Exec("insert into t1 values (500), (-500), (~0), (-1);")
-	terr := errors.Cause(err).(*terror.Error)
-	c.Assert(terr.Code(), Equals, terror.ErrCode(tmysql.ErrWarnDataOutOfRange))
-	s.tk.MustExec("drop table t1")
-
-	s.tk.MustExec("set sql_mode=''")
-	s.tk.MustExec("create table t1 (quantity decimal(2) unsigned);")
-	s.tk.MustExec("insert into t1 values (500), (-500), (~0), (-1);")
-	s.tk.MustQuery("select * from t1").Check(testkit.Rows("99", "0", "99", "0"))
-	s.tk.MustExec("drop table t1")
-}
-
 func (s *testDBSuite) TestCreateIndexType(c *C) {
 	s.tk = testkit.NewTestKit(c, s.store)
 	s.tk.MustExec("use " + s.schemaName)
@@ -947,15 +912,6 @@ func (s *testDBSuite) TestCreateIndexType(c *C) {
 		petsyes enum('Y','N') DEFAULT 'Y' NOT NULL,
 		KEY price (price,area,type,transityes,shopsyes,schoolsyes,petsyes));`
 	s.tk.MustExec(sql)
-}
-
-func (s *testDBSuite) TestIssue3833(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use " + s.schemaName)
-	s.tk.MustExec("create table issue3833 (b char(0))")
-	s.testErrorCode(c, "create index idx on issue3833 (b)", tmysql.ErrWrongKeyColumn)
-	s.testErrorCode(c, "alter table issue3833 add index idx (b)", tmysql.ErrWrongKeyColumn)
-	s.testErrorCode(c, "create table issue3833_2 (b char(0), index (b))", tmysql.ErrWrongKeyColumn)
 }
 
 func (s *testDBSuite) TestColumn(c *C) {
@@ -1688,233 +1644,6 @@ func (s *testDBSuite) TestBitDefaultValue(c *C) {
 	);`)
 }
 
-func (s *testDBSuite) TestCreateTableWithPartition(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use test;")
-	s.tk.MustExec("drop table if exists tp;")
-	s.tk.MustExec(`CREATE TABLE tp (a int) PARTITION BY RANGE(a) (
-	PARTITION p0 VALUES LESS THAN (10),
-	PARTITION p1 VALUES LESS THAN (20),
-	PARTITION p2 VALUES LESS THAN (MAXVALUE)
-	);`)
-	ctx := s.tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(ctx).InfoSchema()
-	tbl, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("tp"))
-	c.Assert(err, IsNil)
-	c.Assert(tbl.Meta().Partition, NotNil)
-	part := tbl.Meta().Partition
-	c.Assert(part.Type, Equals, model.PartitionTypeRange)
-	c.Assert(part.Expr, Equals, "`a`")
-	for _, pdef := range part.Definitions {
-		c.Assert(pdef.ID, Greater, int64(0))
-	}
-	c.Assert(part.Definitions, HasLen, 3)
-	c.Assert(part.Definitions[0].LessThan[0], Equals, "10")
-	c.Assert(part.Definitions[0].Name.L, Equals, "p0")
-	c.Assert(part.Definitions[1].LessThan[0], Equals, "20")
-	c.Assert(part.Definitions[1].Name.L, Equals, "p1")
-	c.Assert(part.Definitions[2].LessThan[0], Equals, "MAXVALUE")
-	c.Assert(part.Definitions[2].Name.L, Equals, "p2")
-
-	s.tk.MustExec("drop table if exists employees;")
-	sql1 := `create table employees (
-	id int not null,
-	hired int not null
-	)
-	partition by range( hired ) (
-		partition p1 values less than (1991),
-		partition p2 values less than (1996),
-		partition p2 values less than (2001)
-	);`
-	s.testErrorCode(c, sql1, tmysql.ErrSameNamePartition)
-
-	sql2 := `create table employees (
-	id int not null,
-	hired int not null
-	)
-	partition by range( hired ) (
-		partition p1 values less than (1998),
-		partition p2 values less than (1996),
-		partition p3 values less than (2001)
-	);`
-	s.testErrorCode(c, sql2, tmysql.ErrRangeNotIncreasing)
-
-	sql3 := `create table employees (
-	id int not null,
-	hired int not null
-	)
-	partition by range( hired ) (
-		partition p1 values less than (1998),
-		partition p2 values less than maxvalue,
-		partition p3 values less than (2001)
-	);`
-	s.testErrorCode(c, sql3, tmysql.ErrPartitionMaxvalue)
-
-	sql4 := `create table t4 (
-	a int not null,
-  	b int not null
-	)
-	partition by range( id ) (
-		partition p1 values less than maxvalue,
-  		partition p2 values less than (1991),
-  		partition p3 values less than (1995)
-	);`
-	s.testErrorCode(c, sql4, tmysql.ErrPartitionMaxvalue)
-
-	_, err = s.tk.Exec(`CREATE TABLE rc (
-    		a INT NOT NULL,
-    		b INT NOT NULL,
-			c INT NOT NULL
-	)
-	partition by range columns(a,b,c) (
-    	partition p0 values less than (10,5,1),
-    	partition p2 values less than (50,maxvalue,10),
-    	partition p3 values less than (65,30,13),
-    	partition p4 values less than (maxvalue,30,40)
-	);`)
-	c.Assert(err, IsNil)
-
-	sql6 := `create table employees (
-	id int not null,
-	hired int not null
-	)
-	partition by range( hired ) (
-		 partition p0 values less than (6 , 10)
-	);`
-	s.testErrorCode(c, sql6, tmysql.ErrTooManyValues)
-
-	sql7 := `create table t7 (
-	a int not null,
-  	b int not null
-	)
-	partition by range( id ) (
-		partition p1 values less than (1991),
-		partition p2 values less than maxvalue,
-  		partition p3 values less than maxvalue,
-  		partition p4 values less than (1995),
-		partition p5 values less than maxvalue
-	);`
-	s.testErrorCode(c, sql7, tmysql.ErrPartitionMaxvalue)
-
-	_, err = s.tk.Exec(`create table t8 (
-	a int not null,
-	b int not null
-	)
-	partition by range( id ) (
-		partition p1 values less than (19xx91),
-		partition p2 values less than maxvalue
-	);`)
-	c.Assert(ddl.ErrNotAllowedTypeInPartition.Equal(err), IsTrue)
-
-	sql9 := `create TABLE t9 (
-	col1 int
-	)
-	partition by range( case when col1 > 0 then 10 else 20 end ) (
-		partition p0 values less than (2),
-		partition p1 values less than (6)
-	);`
-	s.testErrorCode(c, sql9, tmysql.ErrPartitionFunctionIsNotAllowed)
-
-	s.testErrorCode(c, `create TABLE t10 (c1 int,c2 int) partition by range(c1 / c2 ) (partition p0 values less than (2));`, tmysql.ErrPartitionFunctionIsNotAllowed)
-
-	s.tk.MustExec(`create TABLE t11 (c1 int,c2 int) partition by range(c1 div c2 ) (partition p0 values less than (2));`)
-	s.tk.MustExec(`create TABLE t12 (c1 int,c2 int) partition by range(c1 + c2 ) (partition p0 values less than (2));`)
-	s.tk.MustExec(`create TABLE t13 (c1 int,c2 int) partition by range(c1 - c2 ) (partition p0 values less than (2));`)
-	s.tk.MustExec(`create TABLE t14 (c1 int,c2 int) partition by range(c1 * c2 ) (partition p0 values less than (2));`)
-	s.tk.MustExec(`create TABLE t15 (c1 int,c2 int) partition by range( abs(c1) ) (partition p0 values less than (2));`)
-	s.tk.MustExec(`create TABLE t16 (c1 int) partition by range( c1) (partition p0 values less than (10));`)
-
-	s.testErrorCode(c, `create TABLE t17 (c1 int,c2 float) partition by range(c1 + c2 ) (partition p0 values less than (2));`, tmysql.ErrPartitionFuncNotAllowed)
-	s.testErrorCode(c, `create TABLE t18 (c1 int,c2 float) partition by range( floor(c2) ) (partition p0 values less than (2));`, tmysql.ErrPartitionFuncNotAllowed)
-	s.tk.MustExec(`create TABLE t19 (c1 int,c2 float) partition by range( floor(c1) ) (partition p0 values less than (2));`)
-
-	s.tk.MustExec(`create TABLE t20 (c1 int,c2 bit(10)) partition by range(c2) (partition p0 values less than (10));`)
-	s.tk.MustExec(`create TABLE t21 (c1 int,c2 year) partition by range( c2 ) (partition p0 values less than (2000));`)
-
-	s.testErrorCode(c, `create TABLE t24 (c1 float) partition by range( c1 ) (partition p0 values less than (2000));`, tmysql.ErrFieldTypeNotAllowedAsPartitionField)
-
-	// test check order. The sql below have 2 problem: 1. ErrFieldTypeNotAllowedAsPartitionField  2. ErrPartitionMaxvalue , mysql will return ErrPartitionMaxvalue.
-	s.testErrorCode(c, `create TABLE t25 (c1 float) partition by range( c1 ) (partition p1 values less than maxvalue,partition p0 values less than (2000));`, tmysql.ErrPartitionMaxvalue)
-
-	// Fix issue 7362.
-	s.tk.MustExec("create table test_partition(id bigint, name varchar(255), primary key(id)) ENGINE=InnoDB DEFAULT CHARSET=utf8 PARTITION BY RANGE  COLUMNS(id) (PARTITION p1 VALUES LESS THAN (10) ENGINE = InnoDB);")
-
-	// 'Less than' in partition expression could be a constant expression, notice that
-	// the SHOW result changed.
-	s.tk.MustExec(`create table t26 (a date)
-			  partition by range(to_seconds(a))(
-			  partition p0 values less than (to_seconds('2004-01-01')),
-			  partition p1 values less than (to_seconds('2005-01-01')));`)
-	s.tk.MustQuery("show create table t26").Check(
-		testkit.Rows("t26 CREATE TABLE `t26` (\n  `a` date DEFAULT NULL\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin\nPARTITION BY RANGE ( to_seconds(`a`) ) (\n  PARTITION p0 VALUES LESS THAN (63240134400),\n  PARTITION p1 VALUES LESS THAN (63271756800)\n)"))
-	s.tk.MustExec(`create table t27 (a bigint unsigned not null) 	
-		  partition by range(a) (
-		  partition p0 values less than (10),
-		  partition p1 values less than (100),
-		  partition p2 values less than (1000),
-		  partition p3 values less than (18446744073709551000),
-		  partition p4 values less than (18446744073709551614)
-		);`)
-	s.tk.MustExec(`create table t28 (a bigint unsigned not null) 	
-		  partition by range(a) (
-		  partition p0 values less than (10),
-		  partition p1 values less than (100),
-		  partition p2 values less than (1000),
-		  partition p3 values less than (18446744073709551000 + 1),
-		  partition p4 values less than (18446744073709551000 + 10)
-		);`)
-
-	// Only range type partition is now supported, range columns partition only implements the parser, so it will not be checked.
-	// So the following SQL statements can be executed successfully.
-	s.tk.MustExec(`create table t29 (
-		a decimal
-	)
-	partition by range columns (a)
-	(partition p0 values less than (0));`)
-}
-
-func (s *testDBSuite) TestCreateTableWithHashPartition(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use test;")
-	s.tk.MustExec("drop table if exists employees;")
-	s.tk.MustExec("set @@session.tidb_enable_table_partition = 1")
-	s.tk.MustExec(`
-	create table employees (
-		id int not null,
-		fname varchar(30),
-		lname varchar(30),
-		hired date not null default '1970-01-01',
-		separated date not null default '9999-12-31',
-		job_code int,
-		store_id int
-	)
-	partition by hash(store_id) partitions 4;`)
-
-	s.tk.MustExec("drop table if exists employees;")
-	s.tk.MustExec(`
-	create table employees (
-		id int not null,
-		fname varchar(30),
-		lname varchar(30),
-		hired date not null default '1970-01-01',
-		separated date not null default '9999-12-31',
-		job_code int,
-		store_id int
-	)
-	partition by hash( year(hired) ) partitions 4;`)
-}
-
-func (s *testDBSuite) TestCreateTableWithKeyPartition(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use test;")
-	s.tk.MustExec("drop table if exists tm1;")
-	s.tk.MustExec(`create table tm1
-	(
-		s1 char(32) primary key
-	)
-	partition by key(s1) partitions 10;`)
-}
-
 func (s *testDBSuite) TestTableDDLWithFloatType(c *C) {
 	s.tk.MustExec("use test")
 	s.tk.MustExec("drop table if exists t")
@@ -2102,48 +1831,6 @@ out:
 	s.tk.MustQuery("select c2, c3 from tnn where c1 = 99").Check(testkit.Rows(expected))
 
 	s.tk.MustExec("drop table tnn")
-}
-
-func (s *testDBSuite) TestIssue2858And2717(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use " + s.schemaName)
-
-	s.tk.MustExec("create table t_issue_2858_bit (a bit(64) default b'0')")
-	s.tk.MustExec("insert into t_issue_2858_bit value ()")
-	s.tk.MustExec(`insert into t_issue_2858_bit values (100), ('10'), ('\0')`)
-	s.tk.MustQuery("select a+0 from t_issue_2858_bit").Check(testkit.Rows("0", "100", "12592", "0"))
-	s.tk.MustExec(`alter table t_issue_2858_bit alter column a set default '\0'`)
-
-	s.tk.MustExec("create table t_issue_2858_hex (a int default 0x123)")
-	s.tk.MustExec("insert into t_issue_2858_hex value ()")
-	s.tk.MustExec("insert into t_issue_2858_hex values (123), (0x321)")
-	s.tk.MustQuery("select a from t_issue_2858_hex").Check(testkit.Rows("291", "123", "801"))
-	s.tk.MustExec(`alter table t_issue_2858_hex alter column a set default 0x321`)
-}
-
-func (s *testDBSuite) TestIssue4432(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use " + s.schemaName)
-
-	s.tk.MustExec("create table tx (col bit(10) default 'a')")
-	s.tk.MustExec("insert into tx value ()")
-	s.tk.MustQuery("select * from tx").Check(testkit.Rows("\x00a"))
-	s.tk.MustExec("drop table tx")
-
-	s.tk.MustExec("create table tx (col bit(10) default 0x61)")
-	s.tk.MustExec("insert into tx value ()")
-	s.tk.MustQuery("select * from tx").Check(testkit.Rows("\x00a"))
-	s.tk.MustExec("drop table tx")
-
-	s.tk.MustExec("create table tx (col bit(10) default 97)")
-	s.tk.MustExec("insert into tx value ()")
-	s.tk.MustQuery("select * from tx").Check(testkit.Rows("\x00a"))
-	s.tk.MustExec("drop table tx")
-
-	s.tk.MustExec("create table tx (col bit(10) default 0b1100001)")
-	s.tk.MustExec("insert into tx value ()")
-	s.tk.MustQuery("select * from tx").Check(testkit.Rows("\x00a"))
-	s.tk.MustExec("drop table tx")
 }
 
 func (s *testDBSuite) TestChangeColumnPosition(c *C) {
@@ -2720,866 +2407,6 @@ func (s *testDBSuite) TestBackwardCompatibility(c *C) {
 	tk.MustExec("admin check index t idx_b")
 }
 
-func (s *testDBSuite) TestAlterTableAddPartition(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use test;")
-	s.tk.MustExec("drop table if exists employees;")
-	s.tk.MustExec(`create table employees (
-	id int not null,
-	hired date not null
-	)
-	partition by range( year(hired) ) (
-		partition p1 values less than (1991),
-		partition p2 values less than (1996),
-		partition p3 values less than (2001)
-	);`)
-	s.tk.MustExec(`alter table employees add partition (
-    partition p4 values less than (2010),
-    partition p5 values less than MAXVALUE
-	);`)
-
-	ctx := s.tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(ctx).InfoSchema()
-	tbl, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("employees"))
-	c.Assert(err, IsNil)
-	c.Assert(tbl.Meta().Partition, NotNil)
-	part := tbl.Meta().Partition
-	c.Assert(part.Type, Equals, model.PartitionTypeRange)
-
-	c.Assert(part.Expr, Equals, "year(`hired`)")
-	c.Assert(part.Definitions, HasLen, 5)
-	c.Assert(part.Definitions[0].LessThan[0], Equals, "1991")
-	c.Assert(part.Definitions[0].Name, Equals, model.NewCIStr("p1"))
-	c.Assert(part.Definitions[1].LessThan[0], Equals, "1996")
-	c.Assert(part.Definitions[1].Name, Equals, model.NewCIStr("p2"))
-	c.Assert(part.Definitions[2].LessThan[0], Equals, "2001")
-	c.Assert(part.Definitions[2].Name, Equals, model.NewCIStr("p3"))
-	c.Assert(part.Definitions[3].LessThan[0], Equals, "2010")
-	c.Assert(part.Definitions[3].Name, Equals, model.NewCIStr("p4"))
-	c.Assert(part.Definitions[4].LessThan[0], Equals, "MAXVALUE")
-	c.Assert(part.Definitions[4].Name, Equals, model.NewCIStr("p5"))
-
-	s.tk.MustExec("drop table if exists table1;")
-	s.tk.MustExec("create table table1(a int)")
-	sql1 := `alter table table1 add partition (
-		partition p1 values less than (2010),
-		partition p2 values less than maxvalue
-	);`
-	s.testErrorCode(c, sql1, tmysql.ErrPartitionMgmtOnNonpartitioned)
-	s.tk.MustExec(`create table table_MustBeDefined (
-	id int not null,
-	hired date not null
-	)
-	partition by range( year(hired) ) (
-		partition p1 values less than (1991),
-		partition p2 values less than (1996),
-		partition p3 values less than (2001)
-	);`)
-	sql2 := "alter table table_MustBeDefined add partition"
-	s.testErrorCode(c, sql2, tmysql.ErrPartitionsMustBeDefined)
-	s.tk.MustExec("drop table if exists table2;")
-	s.tk.MustExec(`create table table2 (
-
-	id int not null,
-	hired date not null
-	)
-	partition by range( year(hired) ) (
-	partition p1 values less than (1991),
-	partition p2 values less than maxvalue
-	);`)
-
-	sql3 := `alter table table2 add partition (
-		partition p3 values less than (2010)
-	);`
-	s.testErrorCode(c, sql3, tmysql.ErrPartitionMaxvalue)
-
-	s.tk.MustExec("drop table if exists table3;")
-	s.tk.MustExec(`create table table3 (
-	id int not null,
-	hired date not null
-	)
-	partition by range( year(hired) ) (
-	partition p1 values less than (1991),
-	partition p2 values less than (2001)
-	);`)
-
-	sql4 := `alter table table3 add partition (
-		partition p3 values less than (1993)
-	);`
-	s.testErrorCode(c, sql4, tmysql.ErrRangeNotIncreasing)
-
-	sql5 := `alter table table3 add partition (
-		partition p1 values less than (1993)
-	);`
-	s.testErrorCode(c, sql5, tmysql.ErrSameNamePartition)
-
-	sql6 := `alter table table3 add partition (
-		partition p1 values less than (1993),
-		partition p1 values less than (1995)
-	);`
-	s.testErrorCode(c, sql6, tmysql.ErrSameNamePartition)
-
-	sql7 := `alter table table3 add partition (
-		partition p4 values less than (1993),
-		partition p1 values less than (1995),
-		partition p5 values less than maxvalue
-	);`
-	s.testErrorCode(c, sql7, tmysql.ErrSameNamePartition)
-}
-
-func (s *testDBSuite) TestAlterTableDropPartition(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use test")
-	s.tk.MustExec("drop table if exists employees")
-	s.tk.MustExec(`create table employees (
-	id int not null,
-	hired int not null
-	)
-	partition by range( hired ) (
-		partition p1 values less than (1991),
-		partition p2 values less than (1996),
-		partition p3 values less than (2001)
-	);`)
-
-	s.tk.MustExec("alter table employees drop partition p3;")
-	ctx := s.tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(ctx).InfoSchema()
-	tbl, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("employees"))
-	c.Assert(err, IsNil)
-	c.Assert(tbl.Meta().GetPartitionInfo(), NotNil)
-	part := tbl.Meta().Partition
-	c.Assert(part.Type, Equals, model.PartitionTypeRange)
-	c.Assert(part.Expr, Equals, "`hired`")
-	c.Assert(part.Definitions, HasLen, 2)
-	c.Assert(part.Definitions[0].LessThan[0], Equals, "1991")
-	c.Assert(part.Definitions[0].Name, Equals, model.NewCIStr("p1"))
-	c.Assert(part.Definitions[1].LessThan[0], Equals, "1996")
-	c.Assert(part.Definitions[1].Name, Equals, model.NewCIStr("p2"))
-
-	s.tk.MustExec("drop table if exists table1;")
-	s.tk.MustExec("create table table1 (a int);")
-	sql1 := "alter table table1 drop partition p10;"
-	s.testErrorCode(c, sql1, tmysql.ErrPartitionMgmtOnNonpartitioned)
-
-	s.tk.MustExec("drop table if exists table2;")
-	s.tk.MustExec(`create table table2 (
-	id int not null,
-	hired date not null
-	)
-	partition by range( year(hired) ) (
-		partition p1 values less than (1991),
-		partition p2 values less than (1996),
-		partition p3 values less than (2001)
-	);`)
-	sql2 := "alter table table2 drop partition p10;"
-	s.testErrorCode(c, sql2, tmysql.ErrDropPartitionNonExistent)
-
-	s.tk.MustExec("drop table if exists table3;")
-	s.tk.MustExec(`create table table3 (
-	id int not null
-	)
-	partition by range( id ) (
-		partition p1 values less than (1991)
-	);`)
-	sql3 := "alter table table3 drop partition p1;"
-	s.testErrorCode(c, sql3, tmysql.ErrDropLastPartition)
-
-	s.tk.MustExec("drop table if exists table4;")
-	s.tk.MustExec(`create table table4 (
-	id int not null
-	)
-	partition by range( id ) (
-		partition p1 values less than (10),
-		partition p2 values less than (20),
-		partition p3 values less than MAXVALUE
-	);`)
-
-	s.tk.MustExec("alter table table4 drop partition p2;")
-	is = domain.GetDomain(ctx).InfoSchema()
-	tbl, err = is.TableByName(model.NewCIStr("test"), model.NewCIStr("table4"))
-	c.Assert(err, IsNil)
-	c.Assert(tbl.Meta().GetPartitionInfo(), NotNil)
-	part = tbl.Meta().Partition
-	c.Assert(part.Type, Equals, model.PartitionTypeRange)
-	c.Assert(part.Expr, Equals, "`id`")
-	c.Assert(part.Definitions, HasLen, 2)
-	c.Assert(part.Definitions[0].LessThan[0], Equals, "10")
-	c.Assert(part.Definitions[0].Name, Equals, model.NewCIStr("p1"))
-	c.Assert(part.Definitions[1].LessThan[0], Equals, "MAXVALUE")
-	c.Assert(part.Definitions[1].Name, Equals, model.NewCIStr("p3"))
-
-	s.tk.MustExec("drop table if exists tr;")
-	s.tk.MustExec(` create table tr(
-		id int, name varchar(50),
-		purchased date
-	)
-	partition by range( year(purchased) ) (
-    	partition p0 values less than (1990),
-    	partition p1 values less than (1995),
-    	partition p2 values less than (2000),
-    	partition p3 values less than (2005),
-    	partition p4 values less than (2010),
-    	partition p5 values less than (2015)
-   	);`)
-	s.tk.MustExec(`INSERT INTO tr VALUES
-	(1, 'desk organiser', '2003-10-15'),
-	(2, 'alarm clock', '1997-11-05'),
-	(3, 'chair', '2009-03-10'),
-	(4, 'bookcase', '1989-01-10'),
-	(5, 'exercise bike', '2014-05-09'),
-	(6, 'sofa', '1987-06-05'),
-	(7, 'espresso maker', '2011-11-22'),
-	(8, 'aquarium', '1992-08-04'),
-	(9, 'study desk', '2006-09-16'),
-	(10, 'lava lamp', '1998-12-25');`)
-	result := s.tk.MustQuery("select * from tr where purchased between '1995-01-01' and '1999-12-31';")
-	result.Check(testkit.Rows(`2 alarm clock 1997-11-05`, `10 lava lamp 1998-12-25`))
-	s.tk.MustExec("alter table tr drop partition p2;")
-	result = s.tk.MustQuery("select * from tr where purchased between '1995-01-01' and '1999-12-31';")
-	result.Check(testkit.Rows())
-
-	result = s.tk.MustQuery("select * from tr where purchased between '2010-01-01' and '2014-12-31';")
-	result.Check(testkit.Rows(`5 exercise bike 2014-05-09`, `7 espresso maker 2011-11-22`))
-	s.tk.MustExec("alter table tr drop partition p5;")
-	result = s.tk.MustQuery("select * from tr where purchased between '2010-01-01' and '2014-12-31';")
-	result.Check(testkit.Rows())
-
-	s.tk.MustExec("alter table tr drop partition p4;")
-	result = s.tk.MustQuery("select * from tr where purchased between '2005-01-01' and '2009-12-31';")
-	result.Check(testkit.Rows())
-
-	s.tk.MustExec("drop table if exists table4;")
-	s.tk.MustExec(`create table table4 (
-		id int not null
-	)
-	partition by range( id ) (
-		partition Par1 values less than (1991),
-		partition pAR2 values less than (1992),
-		partition Par3 values less than (1995),
-		partition PaR5 values less than (1996)
-	);`)
-	s.tk.MustExec("alter table table4 drop partition Par2;")
-	s.tk.MustExec("alter table table4 drop partition PAR5;")
-	sql4 := "alter table table4 drop partition PAR0;"
-	s.testErrorCode(c, sql4, tmysql.ErrDropPartitionNonExistent)
-}
-
-func (s *testDBSuite) TestAddPartitionTooManyPartitions(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use test")
-	count := ddl.PartitionCountLimit
-	s.tk.MustExec("drop table if exists p1;")
-	sql1 := `create table p1 (
-		id int not null
-	)
-	partition by range( id ) (`
-	for i := 1; i <= count; i++ {
-		sql1 += fmt.Sprintf("partition p%d values less than (%d),", i, i)
-	}
-	sql1 += "partition p1025 values less than (1025) );"
-	s.testErrorCode(c, sql1, tmysql.ErrTooManyPartitions)
-
-	s.tk.MustExec("drop table if exists p2;")
-	sql2 := `create table p2 (
-		id int not null
-	)
-	partition by range( id ) (`
-	for i := 1; i < count; i++ {
-		sql2 += fmt.Sprintf("partition p%d values less than (%d),", i, i)
-	}
-	sql2 += "partition p1024 values less than (1024) );"
-
-	s.tk.MustExec(sql2)
-	sql3 := `alter table p2 add partition (
-   	partition p1025 values less than (1025)
-	);`
-	s.testErrorCode(c, sql3, tmysql.ErrTooManyPartitions)
-}
-
-func checkPartitionDelRangeDone(c *C, s *testDBSuite, partitionPrefix kv.Key) bool {
-	hasOldPartitionData := true
-	for i := 0; i < waitForCleanDataRound; i++ {
-		err := kv.RunInNewTxn(s.store, false, func(txn kv.Transaction) error {
-			it, err := txn.Iter(partitionPrefix, nil)
-			if err != nil {
-				return err
-			}
-			if !it.Valid() {
-				hasOldPartitionData = false
-			} else {
-				hasOldPartitionData = it.Key().HasPrefix(partitionPrefix)
-			}
-			it.Close()
-			return nil
-		})
-		c.Assert(err, IsNil)
-		if !hasOldPartitionData {
-			break
-		}
-		time.Sleep(waitForCleanDataInterval)
-	}
-	return hasOldPartitionData
-}
-
-func (s *testDBSuite) TestTruncatePartitionAndDropTable(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use test;")
-	// Test truncate common table.
-	s.tk.MustExec("drop table if exists t1;")
-	s.tk.MustExec("create table t1 (id int(11));")
-	for i := 0; i < 100; i++ {
-		s.mustExec(c, "insert into t1 values (?)", i)
-	}
-	result := s.tk.MustQuery("select count(*) from t1;")
-	result.Check(testkit.Rows("100"))
-	s.tk.MustExec("truncate table t1;")
-	result = s.tk.MustQuery("select count(*) from t1")
-	result.Check(testkit.Rows("0"))
-
-	// Test drop common table.
-	s.tk.MustExec("drop table if exists t2;")
-	s.tk.MustExec("create table t2 (id int(11));")
-	for i := 0; i < 100; i++ {
-		s.mustExec(c, "insert into t2 values (?)", i)
-	}
-	result = s.tk.MustQuery("select count(*) from t2;")
-	result.Check(testkit.Rows("100"))
-	s.tk.MustExec("drop table t2;")
-	s.testErrorCode(c, "select * from t2;", tmysql.ErrNoSuchTable)
-
-	// Test truncate table partition.
-	s.tk.MustExec("drop table if exists t3;")
-	s.tk.MustExec(`create table t3(
-		id int, name varchar(50),
-		purchased date
-	)
-	partition by range( year(purchased) ) (
-    	partition p0 values less than (1990),
-    	partition p1 values less than (1995),
-    	partition p2 values less than (2000),
-    	partition p3 values less than (2005),
-    	partition p4 values less than (2010),
-    	partition p5 values less than (2015)
-   	);`)
-	s.tk.MustExec(`insert into t3 values
-	(1, 'desk organiser', '2003-10-15'),
-	(2, 'alarm clock', '1997-11-05'),
-	(3, 'chair', '2009-03-10'),
-	(4, 'bookcase', '1989-01-10'),
-	(5, 'exercise bike', '2014-05-09'),
-	(6, 'sofa', '1987-06-05'),
-	(7, 'espresso maker', '2011-11-22'),
-	(8, 'aquarium', '1992-08-04'),
-	(9, 'study desk', '2006-09-16'),
-	(10, 'lava lamp', '1998-12-25');`)
-	result = s.tk.MustQuery("select count(*) from t3;")
-	result.Check(testkit.Rows("10"))
-	ctx := s.tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(ctx).InfoSchema()
-	oldTblInfo, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("t3"))
-	c.Assert(err, IsNil)
-	// Only one partition id test is taken here.
-	oldPID := oldTblInfo.Meta().Partition.Definitions[0].ID
-	s.tk.MustExec("truncate table t3;")
-	partitionPrefix := tablecodec.EncodeTablePrefix(oldPID)
-	hasOldPartitionData := checkPartitionDelRangeDone(c, s, partitionPrefix)
-	c.Assert(hasOldPartitionData, IsFalse)
-
-	// Test drop table partition.
-	s.tk.MustExec("drop table if exists t4;")
-	s.tk.MustExec(`create table t4(
-		id int, name varchar(50),
-		purchased date
-	)
-	partition by range( year(purchased) ) (
-    	partition p0 values less than (1990),
-    	partition p1 values less than (1995),
-    	partition p2 values less than (2000),
-    	partition p3 values less than (2005),
-    	partition p4 values less than (2010),
-    	partition p5 values less than (2015)
-   	);`)
-	s.tk.MustExec(`insert into t4 values
-	(1, 'desk organiser', '2003-10-15'),
-	(2, 'alarm clock', '1997-11-05'),
-	(3, 'chair', '2009-03-10'),
-	(4, 'bookcase', '1989-01-10'),
-	(5, 'exercise bike', '2014-05-09'),
-	(6, 'sofa', '1987-06-05'),
-	(7, 'espresso maker', '2011-11-22'),
-	(8, 'aquarium', '1992-08-04'),
-	(9, 'study desk', '2006-09-16'),
-	(10, 'lava lamp', '1998-12-25');`)
-	result = s.tk.MustQuery("select count(*) from t4; ")
-	result.Check(testkit.Rows("10"))
-	is = domain.GetDomain(ctx).InfoSchema()
-	oldTblInfo, err = is.TableByName(model.NewCIStr("test"), model.NewCIStr("t4"))
-	c.Assert(err, IsNil)
-	// Only one partition id test is taken here.
-	oldPID = oldTblInfo.Meta().Partition.Definitions[1].ID
-	s.tk.MustExec("drop table t4;")
-	partitionPrefix = tablecodec.EncodeTablePrefix(oldPID)
-	hasOldPartitionData = checkPartitionDelRangeDone(c, s, partitionPrefix)
-	c.Assert(hasOldPartitionData, IsFalse)
-	s.testErrorCode(c, "select * from t4;", tmysql.ErrNoSuchTable)
-
-	// Test truncate table partition reassigns new partitionIDs.
-	s.tk.MustExec("drop table if exists t5;")
-	s.tk.MustExec("set @@session.tidb_enable_table_partition=1;")
-	s.tk.MustExec(`create table t5(
-		id int, name varchar(50), 
-		purchased date
-	)
-	partition by range( year(purchased) ) (
-    	partition p0 values less than (1990),
-    	partition p1 values less than (1995),
-    	partition p2 values less than (2000),
-    	partition p3 values less than (2005),
-    	partition p4 values less than (2010),
-    	partition p5 values less than (2015)
-   	);`)
-	is = domain.GetDomain(ctx).InfoSchema()
-	oldTblInfo, err = is.TableByName(model.NewCIStr("test"), model.NewCIStr("t5"))
-	c.Assert(err, IsNil)
-	oldPID = oldTblInfo.Meta().Partition.Definitions[0].ID
-
-	s.tk.MustExec("truncate table t5;")
-	is = domain.GetDomain(ctx).InfoSchema()
-	newTblInfo, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("t5"))
-	c.Assert(err, IsNil)
-	newPID := newTblInfo.Meta().Partition.Definitions[0].ID
-	c.Assert(oldPID != newPID, IsTrue)
-
-	s.tk.MustExec("set @@session.tidb_enable_table_partition = 1;")
-	s.tk.MustExec("drop table if exists clients;")
-	s.tk.MustExec(`create table clients (
-		id int,
-		fname varchar(30),
-		lname varchar(30),
-		signed date
-	)
-	partition by hash( month(signed) )
-	partitions 12;`)
-	is = domain.GetDomain(ctx).InfoSchema()
-	oldTblInfo, err = is.TableByName(model.NewCIStr("test"), model.NewCIStr("clients"))
-	c.Assert(err, IsNil)
-	oldDefs := oldTblInfo.Meta().Partition.Definitions
-
-	// Test truncate `hash partitioned table` reassigns new partitionIDs.
-	s.tk.MustExec("truncate table clients;")
-	is = domain.GetDomain(ctx).InfoSchema()
-	newTblInfo, err = is.TableByName(model.NewCIStr("test"), model.NewCIStr("clients"))
-	c.Assert(err, IsNil)
-	newDefs := newTblInfo.Meta().Partition.Definitions
-	for i := 0; i < len(oldDefs); i++ {
-		c.Assert(oldDefs[i].ID != newDefs[i].ID, IsTrue)
-	}
-}
-
-func (s *testDBSuite) TestPartitionUniqueKeyNeedAllFieldsInPf(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("use test;")
-	s.tk.MustExec("drop table if exists part1;")
-	s.tk.MustExec(`create table part1 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		unique key (col1, col2)
-	)
-	partition by range( col1 + col2 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`)
-
-	s.tk.MustExec("drop table if exists part2;")
-	s.tk.MustExec(`create table part2 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		unique key (col1, col2, col3),
-  		unique key (col3)
-	)
-	partition by range( col3 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`)
-
-	s.tk.MustExec("drop table if exists part3;")
-	s.tk.MustExec(`create table part3 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		primary key(col1, col2)
-	)
-	partition by range( col1 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`)
-
-	s.tk.MustExec("drop table if exists part4;")
-	s.tk.MustExec(`create table part4 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		primary key(col1, col2),
-		unique key(col2)
-	)
-	partition by range( year(col2)  ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`)
-
-	s.tk.MustExec("drop table if exists part5;")
-	s.tk.MustExec(`create table part5 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		primary key(col1, col2, col4),
-    	unique key(col2, col1)
-	)
-	partition by range( col1 + col2 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`)
-
-	s.tk.MustExec("drop table if exists Part1;")
-	sql1 := `create table Part1 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		unique key (col1, col2)
-	)
-	partition by range( col3 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`
-	s.testErrorCode(c, sql1, tmysql.ErrUniqueKeyNeedAllFieldsInPf)
-
-	s.tk.MustExec("drop table if exists Part1;")
-	sql2 := `create table Part1 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		unique key (col1),
-		unique key (col3)
-	)
-	partition by range( col1 + col3 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`
-	s.testErrorCode(c, sql2, tmysql.ErrUniqueKeyNeedAllFieldsInPf)
-
-	s.tk.MustExec("drop table if exists Part1;")
-	sql3 := `create table Part1 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		unique key (col1),
-		unique key (col3)
-	)
-	partition by range( col3 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`
-	s.testErrorCode(c, sql3, tmysql.ErrUniqueKeyNeedAllFieldsInPf)
-
-	s.tk.MustExec("drop table if exists Part1;")
-	sql4 := `create table Part1 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		unique key (col1, col2, col3),
-  		unique key (col3)
-	)
-	partition by range( col1 + col3 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`
-	s.testErrorCode(c, sql4, tmysql.ErrUniqueKeyNeedAllFieldsInPf)
-
-	s.tk.MustExec("drop table if exists Part1;")
-	sql5 := `create table Part1 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		primary key(col1, col2)
-	)
-	partition by range( col3 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`
-	s.testErrorCode(c, sql5, tmysql.ErrUniqueKeyNeedAllFieldsInPf)
-
-	s.tk.MustExec("drop table if exists Part1;")
-	sql6 := `create table Part1 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		primary key(col1, col3),
-		unique key(col2)
-	)
-	partition by range( year(col2)  ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`
-	s.testErrorCode(c, sql6, tmysql.ErrUniqueKeyNeedAllFieldsInPf)
-
-	s.tk.MustExec("drop table if exists Part1;")
-	sql7 := `create table Part1 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		primary key(col1, col3, col4),
-    	unique key(col2, col1)
-	)
-	partition by range( col1 + col2 ) (
-	partition p1 values less than (11),
-	partition p2 values less than (15)
-	);`
-	s.testErrorCode(c, sql7, tmysql.ErrUniqueKeyNeedAllFieldsInPf)
-
-	s.tk.MustExec("drop table if exists part6;")
-	sql8 := `create table part6 (
-		col1 int not null,
-		col2 date not null,
-		col3 int not null,
-		col4 int not null,
-		col5 int not null,
-		unique key(col1, col2),
-		unique key(col1, col3)
-	)
-	partition by range( col1 + col2 ) (
-	partition p1 values less than (11),
-  	partition p2 values less than (15)
-	);`
-	s.testErrorCode(c, sql8, tmysql.ErrUniqueKeyNeedAllFieldsInPf)
-}
-
-func (s *testDBSuite) TestPartitionDropIndex(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	done := make(chan error, 1)
-	s.tk.MustExec("use " + s.schemaName)
-	s.tk.MustExec("drop table if exists partition_drop_idx;")
-	s.tk.MustExec(`create table partition_drop_idx (
-		c1 int, c2 int, c3 int
-	)
-	partition by range( c1 ) (
-    	partition p0 values less than (3),
-    	partition p1 values less than (5),
-    	partition p2 values less than (7),
-    	partition p3 values less than (11),
-    	partition p4 values less than (15),
-    	partition p5 values less than (20),
-		partition p6 values less than (maxvalue)
-   	);`)
-
-	num := 20
-	for i := 0; i < num; i++ {
-		s.mustExec(c, "insert into partition_drop_idx values (?, ?, ?)", i, i, i)
-	}
-	s.tk.MustExec("alter table partition_drop_idx add index idx1 (c1)")
-
-	ctx := s.tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(ctx).InfoSchema()
-	t, err := is.TableByName(model.NewCIStr(s.schemaName), model.NewCIStr("partition_drop_idx"))
-	c.Assert(err, IsNil)
-
-	var idx1 table.Index
-	for _, pidx := range t.Indices() {
-		if pidx.Meta().Name.L == "idx1" {
-			idx1 = pidx
-			break
-		}
-	}
-	c.Assert(idx1, NotNil)
-
-	sessionExecInGoroutine(c, s.store, "drop index idx1 on partition_drop_idx;", done)
-	ticker := time.NewTicker(s.lease / 2)
-	defer ticker.Stop()
-LOOP:
-	for {
-		select {
-		case err = <-done:
-			if err == nil {
-				break LOOP
-			}
-			c.Assert(err, IsNil, Commentf("err:%v", errors.ErrorStack(err)))
-		case <-ticker.C:
-			step := 10
-			rand.Seed(time.Now().Unix())
-			for i := num; i < num+step; i++ {
-				n := rand.Intn(num)
-				s.mustExec(c, "update partition_drop_idx set c2 = 1 where c1 = ?", n)
-				s.mustExec(c, "insert into partition_drop_idx values (?, ?, ?)", i, i, i)
-			}
-			num += step
-		}
-	}
-
-	is = domain.GetDomain(ctx).InfoSchema()
-	t, err = is.TableByName(model.NewCIStr(s.schemaName), model.NewCIStr("partition_drop_idx"))
-	c.Assert(err, IsNil)
-	// Only one partition id test is taken here.
-	pid := t.Meta().Partition.Definitions[0].ID
-	var idxn table.Index
-	t.Indices()
-	for _, idx := range t.Indices() {
-		if idx.Meta().Name.L == "idx1" {
-			idxn = idx
-			break
-		}
-	}
-	c.Assert(idxn, IsNil)
-	idx := tables.NewIndex(pid, t.Meta(), idx1.Meta())
-	checkDelRangeDone(c, ctx, idx)
-	s.tk.MustExec("drop table partition_drop_idx;")
-}
-
-func (s *testDBSuite) TestPartitionCancelAddIndex(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.mustExec(c, "use test_db")
-	s.mustExec(c, "drop table if exists t1;")
-	s.mustExec(c, `create table t1 (
-		c1 int, c2 int, c3 int
-	)
-	partition by range( c1 ) (
-    	partition p0 values less than (1024),
-    	partition p1 values less than (2048),
-    	partition p2 values less than (3072),
-    	partition p3 values less than (4096),
-		partition p4 values less than (maxvalue)
-   	);`)
-	base := defaultBatchSize * 2
-	count := base
-	// add some rows
-	for i := 0; i < count; i++ {
-		s.mustExec(c, "insert into t1 values (?, ?, ?)", i, i, i)
-	}
-
-	var checkErr error
-	var c3IdxInfo *model.IndexInfo
-	hook := &ddl.TestDDLCallback{}
-	oldReorgWaitTimeout := ddl.ReorgWaitTimeout
-	ddl.ReorgWaitTimeout = 10 * time.Millisecond
-	hook.OnJobUpdatedExported, c3IdxInfo, checkErr = backgroundExecOnJobUpdatedExported(c, s, hook)
-	s.dom.DDL().(ddl.DDLForTest).SetHook(hook)
-	done := make(chan error, 1)
-	go backgroundExec(s.store, "create index c3_index on t1 (c3)", done)
-
-	times := 0
-	ticker := time.NewTicker(s.lease / 2)
-	defer ticker.Stop()
-LOOP:
-	for {
-		select {
-		case err := <-done:
-			c.Assert(checkErr, IsNil)
-			c.Assert(err, NotNil)
-			c.Assert(err.Error(), Equals, "[ddl:12]cancelled DDL job")
-			break LOOP
-		case <-ticker.C:
-			if times >= 10 {
-				break
-			}
-			step := 10
-			rand.Seed(time.Now().Unix())
-			// delete some rows, and add some data
-			for i := count; i < count+step; i++ {
-				n := rand.Intn(count)
-				s.mustExec(c, "delete from t1 where c1 = ?", n)
-				s.mustExec(c, "insert into t1 values (?, ?, ?)", i+10, i, i)
-			}
-			count += step
-			times++
-		}
-	}
-
-	t := s.testGetTable(c, "t1")
-	// Only one partition id test is taken here.
-	pid := t.Meta().Partition.Definitions[0].ID
-	for _, tidx := range t.Indices() {
-		c.Assert(strings.EqualFold(tidx.Meta().Name.L, "c3_index"), IsFalse)
-	}
-
-	ctx := s.s.(sessionctx.Context)
-	idx := tables.NewIndex(pid, t.Meta(), c3IdxInfo)
-	checkDelRangeDone(c, ctx, idx)
-
-	s.mustExec(c, "drop table t1")
-	ddl.ReorgWaitTimeout = oldReorgWaitTimeout
-	callback := &ddl.TestDDLCallback{}
-	s.dom.DDL().(ddl.DDLForTest).SetHook(callback)
-}
-
-func backgroundExecOnJobUpdatedExported(c *C, s *testDBSuite, hook *ddl.TestDDLCallback) (func(*model.Job), *model.IndexInfo, error) {
-	var checkErr error
-	first := true
-	ddl.ReorgWaitTimeout = 10 * time.Millisecond
-	c3IdxInfo := &model.IndexInfo{}
-	hook.OnJobUpdatedExported = func(job *model.Job) {
-		addIndexNotFirstReorg := job.Type == model.ActionAddIndex && job.SchemaState == model.StateWriteReorganization && job.SnapshotVer != 0
-		// If the action is adding index and the state is writing reorganization, it want to test the case of cancelling the job when backfilling indexes.
-		// When the job satisfies this case of addIndexNotFirstReorg, the worker will start to backfill indexes.
-		if !addIndexNotFirstReorg {
-			// Get the index's meta.
-			if c3IdxInfo != nil {
-				return
-			}
-			t := s.testGetTable(c, "t1")
-			for _, index := range t.WritableIndices() {
-				if index.Meta().Name.L == "c3_index" {
-					c3IdxInfo = index.Meta()
-				}
-			}
-			return
-		}
-		// The job satisfies the case of addIndexNotFirst for the first time, the worker hasn't finished a batch of backfill indexes.
-		if first {
-			first = false
-			return
-		}
-		if checkErr != nil {
-			return
-		}
-		hookCtx := mock.NewContext()
-		hookCtx.Store = s.store
-		var err error
-		err = hookCtx.NewTxn()
-		if err != nil {
-			checkErr = errors.Trace(err)
-			return
-		}
-		jobIDs := []int64{job.ID}
-		errs, err := admin.CancelJobs(hookCtx.Txn(true), jobIDs)
-		if err != nil {
-			checkErr = errors.Trace(err)
-			return
-		}
-		// It only tests cancel one DDL job.
-		if errs[0] != nil {
-			checkErr = errors.Trace(errs[0])
-			return
-		}
-		err = hookCtx.Txn(true).Commit(context.Background())
-		if err != nil {
-			checkErr = errors.Trace(err)
-		}
-	}
-	return hook.OnJobUpdatedExported, c3IdxInfo, checkErr
-}
-
 func (s *testDBSuite) TestColumnModifyingDefinition(c *C) {
 	s.tk = testkit.NewTestKit(c, s.store)
 	s.tk.MustExec("use test")
@@ -3684,163 +2511,4 @@ LOOP:
 	c.Assert(mysql.HasNotNullFlag(c2.Flag), IsFalse)
 	s.mustExec(c, "drop table t1")
 	ddl.ReorgWaitTimeout = oldReorgWaitTimeout
-}
-
-func (s *testDBSuite) TestPartitionAddIndex(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec(`create table partition_add_idx (
-	id int not null,
-	hired date not null
-	)
-	partition by range( year(hired) ) (
-	partition p1 values less than (1991),
-	partition p3 values less than (2001),
-	partition p4 values less than (2004),
-	partition p5 values less than (2008),
-	partition p6 values less than (2012),
-	partition p7 values less than (2018)
-	);`)
-	testPartitionAddIndex(tk, c)
-
-	// test hash partition table.
-	tk.MustExec("set @@session.tidb_enable_table_partition = '1';")
-	tk.MustExec("drop table if exists partition_add_idx")
-	tk.MustExec(`create table partition_add_idx (
-	id int not null,
-	hired date not null
-	) partition by hash( year(hired) ) partitions 4;`)
-	testPartitionAddIndex(tk, c)
-}
-
-func testPartitionAddIndex(tk *testkit.TestKit, c *C) {
-	for i := 0; i < 500; i++ {
-		tk.MustExec(fmt.Sprintf("insert into partition_add_idx values (%d, '%d-01-01')", i, 1988+rand.Intn(30)))
-	}
-
-	tk.MustExec("alter table partition_add_idx add index idx1 (hired)")
-	tk.MustExec("alter table partition_add_idx add index idx2 (id, hired)")
-	ctx := tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(ctx).InfoSchema()
-	t, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("partition_add_idx"))
-	c.Assert(err, IsNil)
-	var idx1 table.Index
-	for _, idx := range t.Indices() {
-		if idx.Meta().Name.L == "idx1" {
-			idx1 = idx
-			break
-		}
-	}
-	c.Assert(idx1, NotNil)
-
-	tk.MustQuery("select count(hired) from partition_add_idx use index(idx1)").Check(testkit.Rows("500"))
-	tk.MustQuery("select count(id) from partition_add_idx use index(idx2)").Check(testkit.Rows("500"))
-
-	tk.MustExec("admin check table partition_add_idx")
-	tk.MustExec("drop table partition_add_idx")
-}
-
-func (s *testDBSuite) TestDropSchemaWithPartitionTable(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	s.tk.MustExec("drop database if exists test_db_with_partition")
-	s.tk.MustExec("create database test_db_with_partition")
-	s.tk.MustExec("use test_db_with_partition")
-	s.tk.MustExec(`create table t_part (a int key)
-		partition by range(a) (
-		partition p0 values less than (10),
-		partition p1 values less than (20)
-		);`)
-	s.tk.MustExec("insert into t_part values (1),(2),(11),(12);")
-	ctx := s.s.(sessionctx.Context)
-	tbl := s.testGetTableByName(c, "test_db_with_partition", "t_part")
-
-	// check records num before drop database.
-	recordsNum := getPartitionTableRecordsNum(c, ctx, tbl.(table.PartitionedTable))
-	c.Assert(recordsNum, Equals, 4)
-
-	s.tk.MustExec("drop database if exists test_db_with_partition")
-
-	// check job args.
-	rs, err := s.tk.Exec("admin show ddl jobs")
-	c.Assert(err, IsNil)
-	rows, err := session.GetRows4Test(context.Background(), s.tk.Se, rs)
-	c.Assert(err, IsNil)
-	row := rows[0]
-	c.Assert(row.GetString(3), Equals, "drop schema")
-	jobID := row.GetInt64(0)
-	kv.RunInNewTxn(s.store, false, func(txn kv.Transaction) error {
-		t := meta.NewMeta(txn)
-		historyJob, err := t.GetHistoryDDLJob(jobID)
-		c.Assert(err, IsNil)
-		var tableIDs []int64
-		err = historyJob.DecodeArgs(&tableIDs)
-		c.Assert(err, IsNil)
-		// There is 2 partitions.
-		c.Assert(len(tableIDs), Equals, 3)
-		return nil
-	})
-
-	// check records num after drop database.
-	recordsNum = getPartitionTableRecordsNum(c, ctx, tbl.(table.PartitionedTable))
-	c.Assert(recordsNum, Equals, 0)
-}
-
-func getPartitionTableRecordsNum(c *C, ctx sessionctx.Context, tbl table.PartitionedTable) int {
-	num := 0
-	info := tbl.Meta().GetPartitionInfo()
-	for _, def := range info.Definitions {
-		pid := def.ID
-		partition := tbl.(table.PartitionedTable).GetPartition(pid)
-		startKey := partition.RecordKey(math.MinInt64)
-		c.Assert(ctx.NewTxn(), IsNil)
-		err := partition.IterRecords(ctx, startKey, partition.Cols(),
-			func(h int64, data []types.Datum, cols []*table.Column) (bool, error) {
-				num++
-				return true, nil
-			})
-		c.Assert(err, IsNil)
-	}
-	return num
-}
-
-func (s *testDBSuite) TestPartitionErrorCode(c *C) {
-	s.tk = testkit.NewTestKit(c, s.store)
-	// add partition
-	s.tk.MustExec("set @@session.tidb_enable_table_partition = 1")
-	s.tk.MustExec("drop database if exists test_db_with_partition")
-	s.tk.MustExec("create database test_db_with_partition")
-	s.tk.MustExec("use test_db_with_partition")
-	s.tk.MustExec(`create table employees (
-		id int not null,
-		fname varchar(30),
-		lname varchar(30),
-		hired date not null default '1970-01-01',
-		separated date not null default '9999-12-31',
-		job_code int,
-		store_id int
-	)
-	partition by hash(store_id)
-	partitions 4;`)
-	_, err := s.tk.Exec("alter table employees add partition partitions 8;")
-	c.Assert(ddl.ErrUnsupportedAddPartition.Equal(err), IsTrue)
-
-	// coalesce partition
-	s.tk.MustExec(`create table clients (
-		id int,
-		fname varchar(30),
-		lname varchar(30),
-		signed date
-	)
-	partition by hash( month(signed) )
-	partitions 12;`)
-	_, err = s.tk.Exec("alter table clients coalesce partition 4;")
-	c.Assert(ddl.ErrUnsupportedCoalescePartition.Equal(err), IsTrue)
-
-	s.tk.MustExec(`create table t_part (a int key)
-		partition by range(a) (
-		partition p0 values less than (10),
-		partition p1 values less than (20)
-		);`)
-	_, err = s.tk.Exec("alter table t_part coalesce partition 4;")
-	c.Assert(ddl.ErrCoalesceOnlyOnHashPartition.Equal(err), IsTrue)
 }
